% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask}
\alias{mask}
\title{2D Sample Domain Mask.}
\format{A data frame with 21,601 rows and 2 variables:
\describe{
  \item{x}{x coordinate in metres}
  \item{y}{y coordinate in metres}
}}
\usage{
data(mask)
}
\description{
A dataset containing 2D (x and y) coordinates of a 1 × 1 metre regular grid
    that defines the sampled domain associated with the `samples` dataset.
}
\keyword{datasets}
