% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonal-2d.R
\name{plotable_tess}
\alias{plotable_tess}
\title{Create a plotable tesselation dataframe from a point pattern.}
\usage{
plotable_tess(points)
}
\arguments{
\item{points}{2D point pattern object created by package `spatstat`.}
}
\value{
A simple features (sf) polygons dataframe.  May be plotted with
    `ggplot2` using `geom_sf`.
}
\description{
Create a plotable tesselation dataframe from a point pattern.
}
\examples{
library(ggplot2)
dec <- polydeclust2d(samples$x, samples$y, mask = mask)
samples_dec <- cbind(samples, dec$weights)
ptess <- plotable_tess(dec$ppp)
ggplot() +
  geom_raster(data = mask, aes(x, y), fill = "lightblue") +
  geom_sf(data = ptess, fill = NA) +
  geom_point(data = samples_dec, aes(x, y, size = weight))
}
