% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonal-2d.R
\name{plotable_tess}
\alias{plotable_tess}
\title{Create a plotable tesselation dataframe from a point pattern.}
\usage{
plotable_tess(ppp, raster_out = FALSE)
}
\arguments{
\item{ppp}{2D point pattern object created by package `spatstat`.}

\item{raster_out}{Boolean (default `FALSE`). Output the tesselation as a point
grid rather than as a path-type dataframe.}
}
\value{
A dataframe with x and y coordinates and group variable for each
    tessellation path. May be plotted with `ggplot2` using `geom_path` (or
    `geom_raster` in the case of `raster_out = TRUE`.
}
\description{
Create a plotable tesselation dataframe from a point pattern.
}
