% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonal-2d.R
\name{point_spacing_2d}
\alias{point_spacing_2d}
\title{Calculate 2D point sample spacing.}
\usage{
point_spacing_2d(points, nk = 4)
}
\arguments{
\item{points}{A point pattern object from package `spatstat`.}

\item{nk}{Scalar number of nearest points to use.}
}
\value{
A dataframe with columns `x`, `y`, and `distance` defining a grid of
    points (as defined by the `points` window). This can be used in `ggplot`
    with `geom_raster` to vizualize point spacing. Distance is the mean
    distance to the `kn` nearest samples converted to a square grid sapcing:
        `distance = 2 * sqrt(mean_distance^2 / 2)``
    If there are less than `kn` points in the point pattern the distance
    is calculated on the available samples.
}
\description{
Calculate 2D point sample spacing.
}
\examples{
library(ggplot2)
dec <- polydeclust2d(samples$x, samples$y, mask = mask)
spacing <- point_spacing_2d(dec$ppp_mask)
ggplot() +
  geom_raster(data = spacing, aes(x, y, fill = distance)) +
  geom_point(data = samples, aes(x, y)) +
  scale_fill_viridis_c()
}
