% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonal-2d.R
\name{polydeclust2d}
\alias{polydeclust2d}
\title{Calculate decluster weights for 2D samples using tessellations.}
\usage{
polydeclust2d(x, y, mask, expand_mask = 0, normalize = TRUE,
  estdomain = TRUE)
}
\arguments{
\item{x}{Numeric vector. x-axis coordinates with length equal to `y`.}

\item{y}{Numeric vector. x-axis coordinates with length equal to `x`.}

\item{mask}{Numeric dataframe. x and y coordinates of grid mask, This
limits the area of tessellation and hence the magnitude of edge weights.}

\item{expand_mask}{Scalar number (default `0`) of units to exapand mask.
Sometimes when samples are on the mask boundary it is necessary to
slightly expand the mask so that edge samples are included in the
weighting process. Usually only a small value (e.g., 1 unit) is needed.}

\item{normalize}{Boolean (default `TRUE`). Normalize the weights so that they
sum to 1. Otherwise weights sum to the area of the mask.}

\item{estdomain}{Boolean (default `TRUE`). Estimate the spatial domain using
a Ripley-Rasson estimate (`spatstat::ripras`) and intersect with the input
mask to create a window for tesselation. This approach may reduce edge
weights where the input mask is extrapolated well beyond the limits of
the data.}
}
\value{
A labelled list with a dataframe of positive weights optionally
   normalised such that the sum of weights is 1. The weights are initially
   calculated as the ratio of the tesellated area of influence of the sample
   divided by the total mask area. Duplicate samples are assigned a zero
   weight. The list also contains the point pattern object.
}
\description{
Calculate decluster weights for 2D samples using tessellations.
}
